+++ src/test/java/edu/mihail/models/KnightTest.java
@@ -1,8 +1,6 @@
-package edu.mihail.characters;
+package edu.mihail.models;
 
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
